[Flexiglass] Shade to Gone: animate Scrim height to fill screen
This change modifies the scrim animation to ensure its height expands to
fill the entire screen by the end of the transition.

This implementation should be considered a temporary fix.
The final height of this composable is managed by several competing
sources, including multiple modifiers and the STL DSL.

Due to this complexity, the animation may need to be revisited later as
part of a larger refactor (b/438706987).

Test: Manually tested in Shade to Gone/Lockscreen
Bug: 436849876
Flag: com.android.systemui.scene_container
Change-Id: Ie5816bc9852d57f1d19489357f510cf315e42b30