Wrap bundle header in a roundable so that it can round with adjacent swipes.
Additionally sets the ExpandableNotificationRow's max radius to a bundle-height-based value when the ENR is a bundle. The previous implementation allows the corners to become additionally rounded when the bundle is expanded because of the increased height of the ENR overall. But the more-rounded top corner meant that the view during swipe would become oblong when the bottom & top radii on the header are different.

Bug: 421614121
Test: manual, NotificationChildrenContainerTest
Flag: com.android.systemui.notification_bundle_ui

Change-Id: I05951ed9de17056cd9e0117eb4fb6d22731abc55