Set consistent default Locale and TimeZone before each test
Most of these tests were failing when I ran this in deviceless mode,
but setting a consistent locale and time zone before each test seems
to make everything pass. This test is flaking in our postsubmits in
a similar way to what I saw deviceless, so this will hopefully fix
it.

Test: ran test locally in device mode with and without the change
Fixes: 438992055
Flag: EXEMPT test-only change
Change-Id: Ief9d22167125a20b714a20f7d6bc56d5968366ad