Allow open intent to enter PiP w/o config-at-end
Up until now we had been using a config-at-end activity
change to figure out if notification tap to enter PiP
is happening in PiP2, but that's not a really reliable way
of filtering this; sometimes the visibility of activity might
lead to the activity not being marked as config-at-end.

That being said, even when we don't have access to config-at-end
change in Shell, we can still run a legacy cross-fade animation
for enter-PiP.

So this CL does the following 2 things:
 1. Use a flag check instead of pipActivityChange to check for
enter PiP and open-intent mixed transition
 2. Allow OPEN, TO_FRONT and CHANGE modes for PiP task to trigger
legacy animation if there is no activity amongst the changes.

Bug: 439761337
Flag: EXEMPT bugfix
Test: repro the steps in BR manually
Change-Id: I6b14d608b4976b35765d20ad280f9dfa3f272113