Rename ScreenCaptureScope to ScreenCaptureUiScope
The chain starting with this change reworks ScreenCapture scope to
ScreenCaptureUi scope and reintroduces the former back. The difference
would be that ScreenCaptureUi lives when the UI is visible and
ScreenCapture lives when the capture is in progress. This gives us a
neat place to store the data necessary throughout the whole screen
capture, not just at the UI.

Flag: com.android.systemui.new_screen_record_toolbar
Bug: 427682447
Test: passes presubmits
Change-Id: I3960255d525e98ac7165f000cdbfe968072cc714