Replace {@hide} with @hide
In most places the `@hide` doctag is treated as a block tag on its own
line. However, there are a few places where it is treated as an inline
doctag, i.e. `{@hide}`. Dealing with both types increases the
complexity of Metalava and makes it slightly slower.

This change replaces the inline doctag form `{@hide}` with a block tag.

Flag: EXEMPT replacing {@hide} with @hide does not change API or behavior
Bug: 438547944
Test: m checkapi
Change-Id: Iff6ee4f229f61935c8afd63af39614c144824727