Launch split in vertical mode for foldables connected to external display.
When foldable pixel phone is connected to external display, the split selection in connected display should launch split pair in vertical mode. This CL updates the splitLayout to the new config when split select is initiated in external display, and fixes the `isLeftRightSplit` calculation for connected display, so that the split selection can launch in vertical mode as expected.

Bug: 438285797
Test: manual && atest WMShellUnitTests
Flag: com.android.window.flags.enable_non_default_display_split

Change-Id: I9a9458e3ba6802f0c476c64e60c7f9ecfa224e39