Add flag to show a slider UI for routing session volume changes
This change is a non-functional refactor that enables the routing
framework to flag that a UI slider should be associated with a given
session update. The purpose of this change is to show the SysUI volume
slider when a mirroring session's volume changes as a result of a
hardware volume key press.

This change only adds the plumbing. The actual functional change to
populate the flag to true comes in a later change.

The approach used in this change mimics the approach used by
AudioManager#FLAG_SHOW_UI and MediaSessionManager#RemoteSessionCallback.

Bug: b/396394220
Test: atest CtsMediaRouterTestCases
Flag: EXEMPT refactor
Change-Id: I594c110e55eab0b77a7a1e685c55023c5704a9da