Write App Lock supported to ApplicationInfo
We can't write App Lock supported state to ApplicationInfo in the same
place we do for App Lock enabled state because it is not user writable
data and because it requires querying to see if the app is headless
using the ComputerEngine. Write that data when we generate the
ApplicationInfo in the Computer Engine.

Flag: android.security.app_lock_apis
Bug: 436380084
Test: atest PackageManagerSettingsTests
Test: atest AppLockPackageHelperTest
Change-Id: Id272ee70c1f0b22451c652bc86abb460f9a776d4