Prevent dreaming when settings disabled
Previously after a reboot, the mDreamsActivateOnSleepSetting value is
set to the default which is true on some devices, leading to the device
trying to dream instead of doze. This change adds a check for dream
settings enabled, which is default to false.

Test: atest PowerManagerServiceTest
Test: no longer see trying to dream from logs after device restart
Fix: 436573884
Flag: android.service.dreams.nap_when_dream_enabled
Change-Id: Id5d1b399f3789b0f8c3286c89b5bc01fc212526c