Only forceWrite userlist if needed
As a hack to work around flagging, we have to update the userlist
outside of the usual versioning upgrade. But we were too aggressive, and
were updating it if no changes were made. This is unnecessary, and
moreover, caused some test failures (that weren't expecting the
upgrade).

Test: com.android.server.pm.UserManagerServiceUserInfoTest
Flag: EXEMPT BUGFIX
Bug: 435283280
Change-Id: If918280acee09e687bd0784ea922aa0ad5f3e0c6