Use SSID field to match for trusted networks
The AdbKeyStore now uses both BSSID and SSID to check if a network is
trusted. We still use BSSID in the check for networks that were added to
the AdbKeyStore before the SSID was stored.

AdbWifiNetworkMonitor uses the SSID instead of the BSSID to deduplicate
network callbacks.

Test: AdbWifiNetworkMonitorTest
Test: AdbDebuggingManagerTest
Bug: 423809120
Fixes: 444348067
Flag: EXEMPT BUGFIX
Change-Id: I8e34e01d8ee8b267baac03e2146368853bde37c5