On exiting PiP reset override to inherit values from parent
Previously, we expicitely set the value to the parent's configuration.
If the parent's values changed (change of display or simply change of
density or resolution) this would be updated. This changes the code to
set the values as undefined in the override, so the value is inherited
from the parent.

Fix: 443262514
Flag: EXEMPT BUGFIX
Test: atest ActivityRecordTests
Test: Manual test
Change-Id: Id4a82e5988be55fbaca2dca1083bdb727ac66b80