Re-introduce explicit allowance of systemui to new concurrent MessageQueue implementation (DeliQueue)
We previously disabled sysui in concurrent MQ due to an issue with sync barriers that we could not reproduce in tests and did not fully understand.
Since DeliQueue has a completely novel approach to sync barriers, it stands to reason that it's not vulnerable to the same issue that was affecting the previous concurrent implementation.
Seeing as such, we get a do-over.
The team will monitor the situation as the flag ramp progresses.

Bug: 438822893
Flag: build.RELEASE_PACKAGE_MESSAGEQUEUE_IMPLEMENTATION
Change-Id: Ia01132d2694df56595e6d3a66ce48cb97a844979