psc: Move ProcessChangeItem to ProcessListInternal
The `ProcessChangeItem` class was an inner class of
`ActivityManagerService` but is primarily used by `ProcessList` to
notify observers of process state changes.

This CL moves the class to `ProcessListInternal` to improve code
organization and reduce coupling with `ActivityManagerService`. All
references have been updated to the new location.

Bug: 442286304
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I1883bcb0cdbf543ac61a6c26ef4e94ab6b5e620e