Fix a performance regression with add/remove bubble fix
This is a lil complicated. When the expanded view is removed for the
last bubble, it never receives a TO_BACK event because of the changes
in changes for b/416655338.

Normally, it'd get a CLOSE event and things would be fine, however,
if that bubble is added back immediately that won't happen. That means
we don't get the TO_FRONT event when the bubble is added back, because
from WM's perspective, that task was never TO_BACK.

The code in bubbles assumes that we'd get that TO_FRONT event (or
task created event) which triggers setting the content as visible.

I originally tried to fix this by removing the expanded view when
the bubble is dragged out (safe because we alpha out the view) -- this
would trigger the TO_BACK event prior to the removal. However, removing
the view causes a performance regression.

To fix this without introducing a performance regression, I am checking
if the content isn't visible at the end of expand/switch animations
and forcing it visible if it's not.

Flag: com.android.wm.shell.fix_bubbles_add_same_bubble_being_removed
Test: manual - repeatedly get/dismiss 2 bubbles and ensure the content
               is always visible when it should be and nothing crashes
Test: atest BubbleStackViewTest
Test: atest --perf systemui-bubble-1-jank-suite    --iter 25    --class android.platform.test.scenario.sysui.bubble.ExpandAndDragToDismissMicrobenchmark    --metric-filter perfetto_ft_systemui-missed_app_frames-mean
      (compare output with/without flag)
Bug: 442895035
Change-Id: I2908cf5d065855d69babf574fcd9036822c6ed33