Don't change the suggestion during connection or displaying an error.
If the user initiated a connection and that connection is in progress
(STATE_CONNECTING) or failed (STATE_CONNECTING_FAILED), don't replace or
clear the suggested device metadata exposed to the UI when the
underlying suggested device information in the Media Router got changed.

This ensures that the suggested device chip UI is stable during the
connection attempt.

Fix: 441546017
Test: atest SuggestedDeviceManagerTest
Flag: EXEMPT BUGFIX
Change-Id: Ic5004743a5907075070ca7316113608244d590c4