Generic implementation for DevicePolicyManagerService.setPolicy
`setPolicy` now has a generic implementation which can
  * Check permissions
  * Check value type
  * Validate values

This is done using a `PolicyHandler` class, which allows these
generic implementations to be overruled for each policy.
This is done by overriding the methods containing the part of the
behavior you want to overrule.

This CL also introduces a `PolicyInformation` class, which contains
information about a specific policy.
Instances of this class will later be code-generated from the
`@PolicyDefinition` annotations on the `PolicyIdentifier`s, however
for this CL it is hand written and committed.

This `PolicyHandler` is still very incomplete, most specifically:
  * Only `enum` values are supported.
  * It has no generic implementation for storing the policy in the
    DevicePolicyEngine.

Bug: 443666781
Test: m
Test: new unittests (atest FrameworksServicesTests_devicepolicy:com.android.server.devicepolicy.PolicyHandlerTest)
Test: existing tests for ScreenCaptureDisabled (atest android.devicepolicy.cts.ScreenCaptureDisabledTest)
Flag: android.app.admin.flags.policy_streamlining
Change-Id: I7e02b26939bbbc08a69324386e7579c186179d06