No-op when deactivating non-existent desk roots instead of throwing
When a display is disconnected while recents was running in that
display, two things happen roughly at the same time:
1) A new core-started transit request is sent to Shell for handling
the disconnection, which usually leads to the desk root getting removed
2) The ongoing recents transition is finished, which can lead to a desk
deactivation if recents had been entered through a desk

(1) and (2) can race, and if (1) is faster (all the way through
RootTaskDesksOrganizer received the onTaskVanished callback), then by
the time (2) attemps to add the deactivation changes for that root, the
root will no longer exist.

Instead of throwing an error, just no-op since the deactivation changes
for this desk aren't needed anyway if it was removed by the
disconnection.

Flag: EXEMPT bugfix
Fix: 427563407
Test: atest RootTaskDesksOrganizerTest
Change-Id: I848a3e678383b621a56d07a0e5cb90c11b5bafb9